--- plugins/formatDate.ts
@@ -1,20 +1,0 @@
-export default defineNuxtPlugin(() => {
-  function format(date: string, short?: boolean): string {
-    //format date to t.m.Y
-    let months: string[] = ["Januar", "Februar", "März", "April", "Mai", "Juni", "Juli", "August", "September", "Oktober", "November", "Dezember"];
-    if (short) {
-      months = ["Jan", "Feb", "Mär", "Apr", "Mai", "Jun", "Jul", "Aug", "Sep", "Okt", "Nov", "Dez"];
-    }
-    const d: Date = new Date(date);
-    const day = d.getDate();
-    const month = d.getMonth();
-    const year = d.getFullYear();
-    return day + "." + months[month] + "." + year;
-  }
-
-  return {
-    provide: {
-      formatDate: (date: string, short?: boolean) => format(date, short),
-    },
-  };
-});
